% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_types.R
\docType{data}
\name{activity_types}
\alias{activity_types}
\title{Qualtrics activity types}
\format{
A tibble with 22 rows and 1 variables:
\describe{
\item{activity}{activity type (location)}
}
}
\source{
A survey of all activities entered as of spring 2024.
}
\usage{
activity_types
}
\description{
A tibble of all possible activity (oftern referred to as
location) types in the data.
}
\note{
Construction of the data is documented in helper_activity_types.R.

Spelling and grammatical errors are reflected in source Qualitrics
survey data and not corrected.
}
\keyword{datasets}
