% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour_types.R
\docType{data}
\name{behaviour_types}
\alias{behaviour_types}
\title{Qualtrics behaviour types}
\format{
A tibble with 22 rows and 3 variables:
\describe{
\item{type}{behaviour type}
\item{behaviour}{observed behaviour}
\item{color}{color for graphics}
}
}
\source{
Anne Kupfer personal communication
}
\usage{
behaviour_types
}
\description{
Broad categories to which observed behaviours are assigned.
}
\note{
Construction of the data is documented in helper_behaviour_types.R.

Spelling and grammatical errors are reflected in source Qualitrics
survey data and not corrected.

The categories presented differ from the categories detailed in the
Qualtrics survey data. For example, the category titled \verb{Inappropriate Behaviors} is recorded as a mix of \code{Self-Regulation} and \code{Antisocial} in Q3
of the Qualtrics survey.
}
\keyword{datasets}
