% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_upload_observations.R
\name{fetch_survey_data}
\alias{fetch_survey_data}
\title{Tools to facilitate uploading to a database formatted observation
data}
\usage{
fetch_survey_data(survey_id, upload = TRUE, connection)
}
\arguments{
\item{survey_id}{(character) id of Qualtrics survey for which to query observations}

\item{upload}{(logical) logical indicating whether to upload formatted observation data to
the database (TRUE, default) or return them in the form of a tibble}

\item{connection}{(character) database connection}
}
\description{
Tools and a workflow for fetching from Qualtrics, formatting,
then uploading to an sqlite database observation data.

import dplyr
importFrom qualtRics fetch_survey
importFrom janitor clean_names
importFrom tidyselect everything starts_with
importFrom DBI dbWriteTable
importFrom tidyr pivot_longer matches
}
