% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_format_surveys.R
\name{format_surveys}
\alias{format_surveys}
\title{Format survey data harvested from Qualtrics into a more readable form
and ready for upload to the cslobsdb database.}
\usage{
format_surveys(surveys)
}
\arguments{
\item{surveys}{(character) Data frame of surveys resulting from a call to
\code{qualtRics::all_surveys()}}
}
\value{
A formatted data frame of Qualtrics survey data
}
\description{
\code{format_surveys} accepts as input a data frame of survey
data downloaded from Qualtrics sensu \code{qualtRics::all_surveys()}. The
query returns all surveys associated with the provided Qualtrics URL many of
which are not relevant. Among surveys filtered, only those that correspond
to an observation sheet are kept, surveys with "copy" or "- 2" in the name
are removed, and a targeted set of surveys of which the date could not be
determined are removed. Year, semester, and class are estimated from file
metadata. Dates are formatted, and field names are cleaned (per janitor).
Although not accessible through the web interface, querying survey data
through the API can return multiple survey instances for a given year,
semester, class, reliability combination. It is not clear the reason for
more than one instance of the same survey, but the observation data are
different among them. Only the most recently modified survey among a
duplicate set is returned.
}
