% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_behaviour_types.R
\name{generate_behaviour_types}
\alias{generate_behaviour_types}
\title{R code to construct behaviour_types package data object}
\usage{
generate_behaviour_types()
}
\description{
R code to construct the \code{behaviour_types} data object in
the cslqualtrics package; included for reference (not exported).

importFrom tibble tribble
}
\examples{
\dontrun{
behaviour_types <- generate_behaviour_types()
usethis::use_data(behaviour_types, overwrite = TRUE)
}

}
