% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_upload_observations.R
\docType{data}
\name{observation_expected_cols}
\alias{observation_expected_cols}
\title{Tools to facilitate uploading to a database formatted observation
data}
\format{
An object of class \code{character} of length 20.
}
\usage{
observation_expected_cols
}
\description{
Tools and a workflow for fetching from Qualtrics, formatting,
then uploading to an sqlite database observation data.

import dplyr
importFrom qualtRics fetch_survey
importFrom janitor clean_names
importFrom tidyselect everything starts_with
importFrom DBI dbWriteTable
importFrom purrr possibly
importFrom tidyr pivot_longer matches
}
\keyword{datasets}
