% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_query_observations.R
\name{query_class_observations}
\alias{query_class_observations}
\title{query a bird observation record}
\usage{
query_class_observations(survey_id)
}
\description{
The function included here ( query_bird_observation ) ...
}
