% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_sql_execution.R
\name{run_interpolated_query}
\alias{run_interpolated_query}
\title{helpers to facilitate sql executions}
\usage{
run_interpolated_query(interpolatedQuery)
}
\description{
The functions run_interpolated_query() and
run_interpolated_execution() are (sub)helper functions that establish and
close database connections to facilitate either a SQL query to return an
objet or execute a statement.
}
