% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{server component of the cslqualtrics Shiny application}
\usage{
server(input, output, session)
}
\description{
The server component of the cslqualtrics Shiny application.
Developed as a single file that is not drawing upon modules.
}
\note{
To facilitate search within the columns of a table but not globally
(i.e., remove the search bar), we set the dom to 'tp'. However, this
approach is deprecated (see: https://datatables.net/reference/option/dom)
and will not be supported as of datatables v3. At the time this app was
developed, the recommended approach was not available in Shiny.

import shiny
import dplyr
import ggplot2
import qualtRics
import pool
importFrom DT renderDT JS
importFrom glue glue_sql
importFrom DBI ANSI dbWriteTable
importFrom purrr walk possibly
importFrom lobstr tree
}
