% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{UI component of the cslqualtrics Shiny application}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
ui
}
\description{
The UI component of the cslqualtrics Shiny application.
Developed as a single UI that is not drawing upon modules.

import shiny
import bslib
importFrom DT DTOutput
}
\keyword{datasets}
