% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_query_surveys.R
\name{query_surveys}
\alias{query_surveys}
\title{Query survey data from sqlite database}
\usage{
query_surveys()
}
\description{
The function included here (\code{query_surveys}) queries all
survey data in the sqlite database. Additionally, the number of observation
records associated with each survey is calculated and returned as part of
the result.

importFrom glue glue_sql
importFrom DBI ANSI
}
